+++ packages/ice/src/bundler/webpack/build.ts
@@ -52,7 +52,10 @@ async function build(
   if (isSuccessful) {
     // Generate html when SSG.
     const outputDir = webpackConfigs[0].output.path;
-    const { serverEntry } = await extendsPluginAPI.serverCompileTask.get() || {};
+    const { serverEntry, error } = await extendsPluginAPI.serverCompileTask.get() || {};
+    if (error) {
+      throw new Error('Build failed, please check the error message.');
+    }
     const outputPaths = await getOutputPaths({
       rootDir,
       serverEntry,