+++ packages/runtime/src/types.ts
@@ -99,6 +99,7 @@ export interface LoaderData {
 export interface AppContext {
   appConfig: AppConfig;
   appData: any;
+  documentData?: any;
   serverData?: any;
   assetsManifest?: AssetsManifest;
   loaderData?: LoadersData;