+++ packages/rspack-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ice/rspack-config
 
+## 1.0.2
+
+### Patch Changes
+
+- e3e1f27c: fix: compatible with win32 when enable --speedup
+- 1fa826b0: chore: bump rspack version for more hook and plugin support
+- Updated dependencies [2875b544]
+- Updated dependencies [101eadea]
+- Updated dependencies [1fa826b0]
+  - @ice/bundles@0.1.15
+  - @ice/shared-config@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes